pyproject.toml
[tool.poetry]
name = "kimaideck"
version = "0.1.0"
description = ""
authors = ["Johann Wagner <johann.wagner@wobcom.de>"]
license = "MIT"
readme = "README.md"

include = ["kimaideck/assets"]

[tool.poetry.dependencies]
python = "^3.11"
streamdeck = "^0.9.5"
pillow = "^10.3.0"
requests = "^2.31.0"
pyyaml = "^6.0.1"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
kimaideck = "kimaideck.__main__:main"